Refactor RSpace exporter extensions to syntax object
